feat: Allow custom disk configurations
This commit introduces the ability to specify disk size and type in the GCENodeClass.

Previously, the disk size and type were hardcoded, ignoring any user-defined values in the GCENodeClass. This change reads the `disk` property from the `GCENodeClassSpec` and applies the specified `sizeGiB` and `categories` to the created instances.

This allows users to define custom disk configurations for their node classes, for example:

```yaml
disk:
  sizeGiB: 256
  categories:
  - "pd-ssd"
```

This change also includes a default disk size of 100GB to prevent any breaking changes.

Fixes: #109
Signed-off-by: cr0wnf0x <zurkuhlf@gmail.com>